[EM-83] correct folders
