--- backend/src/main/java/it/andmora/backend/ExpensesMonitorApplication.java
@@ -1,13 +1,0 @@
-package it.andmora.backend;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ExpensesMonitorApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ExpensesMonitorApplication.class, args);
-	}
-
-}
