--- backend/src/main/java/it/andmora/backend/dao/dbmodel/PaymentTagsDbEntity.java
@@ -1,18 +1,0 @@
-package it.andmora.backend.dao.dbmodel;
-
-import lombok.Builder;
-import lombok.Data;
-import org.springframework.data.annotation.Transient;
-import org.springframework.data.relational.core.mapping.Table;
-
-@Table("payment_tags")
-@Builder
-@Data
-public class PaymentTagsDbEntity {
-
-  @Transient
-  private TagDbEntity tag;
-  @Transient
-  private PaymentDbEntity payment;
-
-}
