--- backend/src/main/java/it/andmora/backend/domain/usecase/BalanceCalculator.java
@@ -1,11 +1,0 @@
-package it.andmora.backend.domain.usecase;
-
-import reactor.core.publisher.Mono;
-
-/**
- * use case to calculate the account balance
- */
-public interface BalanceCalculator {
-
-  Mono<Integer> getOverallBalance();
-}
