--- backend/src/main/java/it/andmora/backend/web/BalanceController.java
@@ -1,12 +1,0 @@
-package it.andmora.backend.web;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import reactor.core.publisher.Mono;
-
-@RequestMapping("balance")
-public interface BalanceController {
-
-  @GetMapping
-  Mono<Integer> getOverallBalance();
-}
