--- backend/src/main/java/it/andmora/backend/web/HomepageController.java
@@ -1,14 +1,0 @@
-package it.andmora.backend.web;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HomepageController {
-
-  @GetMapping
-  String homepage() {
-    return "Hello";
-  }
-
-}
