+++ backend/src/main/java/it/andmora/expensesmonitor/backend/ExpensesMonitorApplication.java
@@ -1,0 +1,13 @@
+package it.andmora.expensesmonitor.backend;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ExpensesMonitorApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ExpensesMonitorApplication.class, args);
+	}
+
+}
