+++ backend/src/main/java/it/andmora/expensesmonitor/backend/domain/PaymentDao.java
@@ -1,0 +1,11 @@
+package it.andmora.expensesmonitor.backend.domain;
+
+import it.andmora.expensesmonitor.backend.domain.model.Payment;
+import reactor.core.publisher.Mono;
+
+public interface PaymentDao {
+
+  Mono<Payment> savePayment(Payment payment);
+
+  Mono<Integer> getOverallBalance();
+}
