+++ backend/src/main/java/it/andmora/expensesmonitor/backend/domain/model/Payment.java
@@ -1,0 +1,21 @@
+package it.andmora.expensesmonitor.backend.domain.model;
+
+import java.time.LocalDateTime;
+import java.util.Set;
+import lombok.Builder;
+import lombok.Data;
+
+/**
+ * Domain object to represent the single payment
+ */
+@Builder
+@Data
+public class Payment {
+
+  private String description;
+  private int amount;
+  private String merchantName;
+  private LocalDateTime accountingDate;
+  private String category;
+  private Set<String> tags;
+}
