+++ backend/src/main/java/it/andmora/expensesmonitor/backend/domain/usecase/PaymentCreator.java
@@ -1,0 +1,12 @@
+package it.andmora.expensesmonitor.backend.domain.usecase;
+
+import it.andmora.expensesmonitor.backend.domain.model.Payment;
+import reactor.core.publisher.Mono;
+
+/**
+ * Use case that implements the creation of a new Payment
+ */
+public interface PaymentCreator {
+
+  Mono<Payment> createPayment(Payment payment);
+}
