+++ backend/src/main/java/it/andmora/expensesmonitor/backend/web/BalanceController.java
@@ -1,0 +1,12 @@
+package it.andmora.expensesmonitor.backend.web;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import reactor.core.publisher.Mono;
+
+@RequestMapping("balance")
+public interface BalanceController {
+
+  @GetMapping
+  Mono<Integer> getOverallBalance();
+}
