+++ backend/src/main/java/it/andmora/expensesmonitor/backend/web/HomepageController.java
@@ -1,0 +1,14 @@
+package it.andmora.expensesmonitor.backend.web;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class HomepageController {
+
+  @GetMapping
+  String homepage() {
+    return "Hello";
+  }
+
+}
