+++ backend/src/main/java/it/andmora/expensesmonitor/backend/web/PaymentController.java
@@ -1,0 +1,14 @@
+package it.andmora.expensesmonitor.backend.web;
+
+import it.andmora.expensesmonitor.backend.web.dto.PaymentDto;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import reactor.core.publisher.Mono;
+
+@RequestMapping("payment")
+public interface PaymentController {
+
+  @PutMapping("")
+  Mono<PaymentDto> createPayment(@RequestBody PaymentDto paymentDto);
+}
