--- backend/src/test/java/it/andmora/expensesmonitor/ExpensesMonitorApplicationTests.java
@@ -1,16 +1,0 @@
-package it.andmora.expensesmonitor;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ExpensesMonitorApplicationTests {
-
-	@Test
-	void contextLoads() {
-		assertThat(true).isTrue();
-	}
-
-}
