+++ backend/src/test/java/it/andmora/expensesmonitor/backend/ExpensesMonitorApplicationTests.java
@@ -1,0 +1,16 @@
+package it.andmora.expensesmonitor.backend;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ExpensesMonitorApplicationTests {
+
+	@Test
+	void contextLoads() {
+		assertThat(true).isTrue();
+	}
+
+}
